fix issue related to chat width stays same on resize
#1491
Previously, on resize, when the view became tablet view we updated chatWidth to maxChatSize and once the window size again became desktop view we either saw a strange behaviour, or the width was defaulted to min chat width. Now we store separately realChatWidth and have separate variable for the chat width which actually is used, so on resizes we have proper width for chat.